requirements.txt
fastapi
click
sqlalchemy
arrow
requests
pydash
alembic
psycopg2
loguru
quantstats
tqdm
pandas
ta_lib_easy
numpy